pubkey_load: replace ARG_CHECK with VERIFY_CHECK
This change prepares for the addition of ge_from_bytes and ge_to_bytes which do
not take a context. This change should be inconsequential because it's very
unlikely to land in the `sizeof(secp256k1_ge_storage) != 64` branch and the
other branch does not have an ARG_CHECK either.